chore: move shared hooks with stubs
