+++ src/hooks/index.ts
@@ -1,3 +1,3 @@
-// Re-export hooks
-export { useApiContent } from './useApiContent';
-
+// TODO: STUB REMOVE BY: 2026-03-31
+// Ticket: PHASE2-002
+export * from '@/shared/hooks/index';
