+++ src/shared/hooks/index.ts
@@ -1,0 +1,2 @@
+// Re-export hooks
+export { useApiContent } from './useApiContent';
